fix/ RemoteComputerManager class created.
